Task 102-type_checking.py

+++ 0x00-python_variable_annotations/102-type_checking.py
@@ -5,7 +5,7 @@ Type Checking Module. Advanced for function zoom Array
 from typing import Sequence, Generator, Any
 
 
-def zoom_array(lst: Sequence[int], factor: float = 2) -> Sequence[Any]:
+def zoom_array(lst: Sequence[int], factor: int = 2) -> Sequence[Any]:
     """Zoom Array function: Zooms an array factor times"""
     zoomed_in: Sequence[Any] = [
         item for item in lst
@@ -18,4 +18,4 @@ array: Sequence[int] = [12, 72, 91]
 
 zoom_2x: Sequence[Any] = zoom_array(array)
 
-zoom_3x: Sequence[Any] = zoom_array(array, 3.0)
+zoom_3x: Sequence[Any] = zoom_array(array, 3)
